MC-2025: Save and load entity AABB to prevent wobble
What follows is a summarized analysis provided on the Mojira and associated subreddit by various MC community members
who have investigated this issue and found a solution. This work is largely a result of their efforts.

The underlying cause of MC-2025 is that sometimes an AABB ends up slightly smaller than the desired width. If this
happens before the entity is pushed up against a boundary (i.e. blocks or walls), then upon chunk save and reload, the
AABB size will be recomputed such that it is intersecting the wall, allowing the entity to be pushed into the wall,
suffocate, and die. Although the rounding artifacts get larger at larger world coordinates, the drift we see is
miniscule. Closer to the world origin, we have seen error on the order of 2-46. Compare this to the fact that
(due to MC-4), entity coordinates send to clients are quantized to multiples of 1/4096 (2-12).

But, OMG, do the rounding errors mean that AABB's accumulate shrinkage over time? Actually, no. The statistics on IEEE
rounding do not have that kind of bias. What has not been stated is that the AABB is just as likely to end up larger
than the expected width; on save and reload, the entity ends up slightly away from the wall, and we don't notice any
problem. In reality, the AABB size ends up undergoing random wobble around the expected value all the time, and that
wobble isn't functionally any different from the kind we'd get even if we tried to force the AABB size to be stable!

This reasoning leads us to one clear conclusion: The simplest, least invasive, and most correct solution is to just
store the AABB in NBT data on chunk save and restore the AABB exactly as it was saved upon reload.